+++ fiddles/angular2-cli/fiddle-0017-SlickGridTreePlus/src/app/components/index.ts
@@ -1,0 +1,2 @@
+export * from './tree-grid/index';
+export * from './shared/index';