+++ fiddles/angular2-cli/fiddle-0017-SlickGridTreePlus/src/app/components/search-bar/search-bar.module.ts
@@ -1,0 +1,18 @@
+import { NgModule } from '@angular/core';
+import { ReactiveFormsModule } from '@angular/forms';
+import { SearchBarComponent } from './index';
+
+@NgModule({
+  imports: [
+    ReactiveFormsModule
+  ],
+  declarations: [
+    SearchBarComponent
+  ],
+  exports: [
+    SearchBarComponent
+  ]
+})
+
+export class SearchBarModule {
+}