+++ fiddles/angular2-cli/fiddle-0017-SlickGridTreePlus/src/app/components/tree-grid/enums/tree-grid-cmds.enum.ts
@@ -1,0 +1,10 @@
+export enum TreeGridCmds {
+  ShowAll = 1,
+  ShowSelected = 2,
+  ClearSelected = 3,
+  Reload = 4,
+  ExpandAll = 5,
+  CollapseAll = 6,
+  SelectAll = 7,
+  Export = 8
+}