+++ fiddles/angular2-cli/fiddle-0017-SlickGridTreePlus/src/app/components/tree-grid/enums/tree-grid-events.enum.ts
@@ -1,0 +1,8 @@
+
+export enum TreeGridEvents {
+  cmd = 1,
+  nodeChange = 2,
+  filter = 3,
+  selectBubble = 4,
+  nodeRequest = 5
+}