+++ fiddles/angular2-cli/fiddle-0017-SlickGridTreePlus/src/app/components/tree-grid/index.ts
@@ -1,0 +1,6 @@
+export * from './enums/index';
+export * from './interfaces/index';
+export * from './tree-grid.component';
+export * from './tree-grid.module';
+export * from './services/index';
+