+++ fiddles/angular2-cli/fiddle-0017-SlickGridTreePlus/src/app/components/tree-grid/tree-grid.module.ts
@@ -1,0 +1,23 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { ReactiveFormsModule } from '@angular/forms';
+import { TreeGridComponent } from './tree-grid.component';
+import { SearchBarModule } from '../search-bar/search-bar.module';
+
+
+@NgModule({
+  imports: [
+    ReactiveFormsModule,
+    SearchBarModule,
+    CommonModule
+  ],
+  declarations: [
+    TreeGridComponent
+  ],
+  exports: [
+    TreeGridComponent
+  ]
+})
+export class TreeGridModule {
+
+}