+++ fiddles/angular2-cli/fiddle-0018-SlickGridSparkLine/e2e/app.e2e-spec.ts
@@ -1,0 +1,14 @@
+import { SlickGridQuickFilterPage } from './app.po';
+
+describe('slick-grid-quick-filter App', () => {
+  let page: SlickGridQuickFilterPage;
+
+  beforeEach(() => {
+    page = new SlickGridQuickFilterPage();
+  });
+
+  it('should display welcome message', () => {
+    page.navigateTo();
+    expect(page.getParagraphText()).toEqual('Welcome to app!');
+  });
+});